data-pipelines/dags/spark-app/requirements.txt
# Spark Batch Job requirements
apache-airflow==2.4.0
pyspark==3.1.2
kubernetes==23.3.0
requests==2.26.0
